requirements.txt
mssql-django>=1.5
pyodbc>=4.0
openpyxl>=3.1
django-htmx>=1.15.0
django-environ>=0.12.0